Add a special user mode
In user mode (when the redaction ID is 0) redact_changeset.rb does not
redact the objects. As an replacement, it writes a file of commands
which call woodpeck's redaction.pl script.
https://github.com/woodpeck/osm-revert-scripts

In some cases you do not want to redact changes because their license is
ok but you want to revert them. Reverts are difficult if later changes
have been applied to the edited objects (you get conflicts). If you can
solve the conflicts automatically (has some downsides), you can revert
them easier. The script calling redaction.pl which is written as an
output, can be used by a DWG member to redact the old versions later if
the need to redact them arises.